Manage multiple CDISCTealDataConnector, CDISCTealDatasetConnector and CDISCTealDataset objects.

Class manages CDISCTealDataConnector, CDISCTealDatasetConnector and CDISCTealDataset objects and aggregate them in one collection.

Super classes

teal.data::TealDataAbstract -> teal.data::TealData -> CDISCTealData

Methods

Public methods

CDISCTealData$new()

CDISCTealData$get_parent()

CDISCTealData$check_metadata()

CDISCTealData$clone()

Inherited methods

teal.data::TealDataAbstract$check()

teal.data::TealDataAbstract$check_reproducibility()

teal.data::TealDataAbstract$execute_mutate()

teal.data::TealDataAbstract$get_check()

teal.data::TealDataAbstract$get_check_result()

teal.data::TealDataAbstract$get_code()

teal.data::TealDataAbstract$get_code_class()

teal.data::TealDataAbstract$get_dataset()

teal.data::TealDataAbstract$get_datasets()

teal.data::TealDataAbstract$is_pulled()

teal.data::TealDataAbstract$mutate()

teal.data::TealDataAbstract$mutate_dataset()

teal.data::TealDataAbstract$reassign_datasets_vars()

teal.data::TealDataAbstract$set_check()

teal.data::TealDataAbstract$set_pull_code()

teal.data::TealData$copy()

teal.data::TealData$get_connectors()

teal.data::TealData$get_datanames()

teal.data::TealData$get_items()

teal.data::TealData$get_join_keys()

teal.data::TealData$get_server()

teal.data::TealData$get_ui()

teal.data::TealData$launch()

teal.data::TealData$mutate_join_keys()

teal.data::TealData$print()

Method new()

Create a new object of CDISCTealData class

Usage

CDISCTealData$new(..., check = FALSE, join_keys)

Arguments

...

(TealDataConnector, TealDataset or TealDatasetConnector) elements to include where ADSL data is mandatory.

check

(logical) reproducibility check - whether evaluated preprocessing code gives the same objects as provided in arguments. Check is run only if flag is true and preprocessing code is not empty.

join_keys

(JoinKeys) or a single (JoinKeySet)
(optional) object with datasets column names used for joining. If empty then it would be automatically derived basing on intersection of datasets primary keys

Method get_parent()

Get all datasets parent names

Usage

CDISCTealData$get_parent()

Returns

(named list) with dataset name and its corresponding parent dataset name

Method check_metadata()

Check correctness of stored joining keys and presence of keys to parent

Usage

CDISCTealData$check_metadata()

Returns

raise and error or invisible TRUE

Method clone()

The objects of this class are cloneable with this method.

Usage

CDISCTealData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.