Manage multiple and TealDatasetConnector of the same type.

Class manages TealDatasetConnector to specify additional dynamic arguments and to open/close connection.

Super classes

teal.data::TealDataAbstract -> teal.data::TealDataConnector -> CDISCTealDataConnector

Methods

Public methods

CDISCTealDataConnector$new()

CDISCTealDataConnector$get_parent()

CDISCTealDataConnector$clone()

Inherited methods

teal.data::TealDataAbstract$check()

teal.data::TealDataAbstract$check_reproducibility()

teal.data::TealDataAbstract$execute_mutate()

teal.data::TealDataAbstract$get_check()

teal.data::TealDataAbstract$get_check_result()

teal.data::TealDataAbstract$get_code()

teal.data::TealDataAbstract$get_datanames()

teal.data::TealDataAbstract$get_dataset()

teal.data::TealDataAbstract$get_datasets()

teal.data::TealDataAbstract$get_items()

teal.data::TealDataAbstract$is_pulled()

teal.data::TealDataAbstract$mutate_dataset()

teal.data::TealDataAbstract$reassign_datasets_vars()

teal.data::TealDataAbstract$set_check()

teal.data::TealDataAbstract$set_pull_code()

teal.data::TealDataConnector$get_code_class()

teal.data::TealDataConnector$get_connection()

teal.data::TealDataConnector$get_preopen_server()

teal.data::TealDataConnector$get_server()

teal.data::TealDataConnector$get_ui()

teal.data::TealDataConnector$is_failed()

teal.data::TealDataConnector$launch()

teal.data::TealDataConnector$mutate()

teal.data::TealDataConnector$print()

teal.data::TealDataConnector$pull()

teal.data::TealDataConnector$set_preopen_server()

teal.data::TealDataConnector$set_pull_args()

teal.data::TealDataConnector$set_server()

teal.data::TealDataConnector$set_ui()

Method new()

Create a new CDISCTealDataConnector object

Usage

CDISCTealDataConnector$new(connection, connectors)

Arguments

connection

(TealDataConnection)
connection to data source

connectors

(list of TealDatasetConnector elements)
list with dataset connectors

Method get_parent()

Get all datasets parent names

Usage

CDISCTealDataConnector$get_parent()

Returns

(named list) with dataset name and its corresponding parent dataset name

Method clone()

The objects of this class are cloneable with this method.

Usage

CDISCTealDataConnector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.