R6 Class representing a MultiAssayExperiment object with its attributes

Any MultiAssayExperiment object can be stored inside this MAETealDataset. Some attributes like colnames, dimension or column names for a specific type will be automatically derived.

See also

Super class

teal.data::TealDataset -> MAETealDataset

Methods

Public methods

MAETealDataset$new()

MAETealDataset$check()

MAETealDataset$check_keys()

MAETealDataset$get_colnames()

MAETealDataset$get_column_labels()

MAETealDataset$get_ncol()

MAETealDataset$get_nrow()

MAETealDataset$get_rownames()

MAETealDataset$print()

MAETealDataset$clone()

Inherited methods

teal.data::TealDataset$get_attrs()

teal.data::TealDataset$get_character_colnames()

teal.data::TealDataset$get_code()

teal.data::TealDataset$get_code_class()

teal.data::TealDataset$get_dataname()

teal.data::TealDataset$get_datanames()

teal.data::TealDataset$get_dataset()

teal.data::TealDataset$get_dataset_label()

teal.data::TealDataset$get_factor_colnames()

teal.data::TealDataset$get_keys()

teal.data::TealDataset$get_metadata()

teal.data::TealDataset$get_mutate_code_class()

teal.data::TealDataset$get_mutate_vars()

teal.data::TealDataset$get_numeric_colnames()

teal.data::TealDataset$get_raw_data()

teal.data::TealDataset$get_row_labels()

teal.data::TealDataset$get_var_r6()

teal.data::TealDataset$get_vars()

teal.data::TealDataset$is_mutate_delayed()

teal.data::TealDataset$is_pulled()

teal.data::TealDataset$mutate()

teal.data::TealDataset$reassign_datasets_vars()

teal.data::TealDataset$recreate()

teal.data::TealDataset$set_code()

teal.data::TealDataset$set_dataset_label()

teal.data::TealDataset$set_keys()

teal.data::TealDataset$set_vars()

Method new()

Create a new object of MAETealDataset class

Usage

MAETealDataset$new(
  dataname,
  x,
  keys = character(0),
  code = character(0),
  label = character(0),
  vars = list(),
  metadata = NULL
)

Arguments

dataname

(character)
A given name for the dataset it may not contain spaces

x

(MultiAssayExperiment)

keys

optional, (character)
A vector of primary keys

code

(character or CodeClass)
A character string defining the code needed to produce the data set in x. initialize()andrecreate()accept code asCodeClass` which is also needed to preserve the code uniqueness and correct order.

label

(character)
Label to describe the dataset

vars

(named list))
In case when this object code depends on other TealDataset object(s) or other constant value, this/these object(s) should be included as named element(s) of the list. For example if this object code needs ADSL object we should specify vars = list(ADSL = <adsl object>). It is recommended to include TealDataset or TealDatasetConnector objects to the vars list to preserve reproducibility. Please note that vars are included to this object as local vars and they cannot be modified within another dataset.

metadata

(named list or NULL)
Field containing metadata about the dataset. Each element of the list should be atomic and of length one.

Method check()

Check to determine if the raw data is reproducible from the get_code() code.

Usage

MAETealDataset$check()

Returns

TRUE if the dataset generated from evaluating the get_code() code is identical to the raw data, else FALSE.

Method check_keys()

Check if keys has been specified correctly for dataset. Set of keys should distinguish unique rows or be character(0).

Usage

MAETealDataset$check_keys(keys = private$.keys)

Arguments

keys

optional, (character)
A vector of primary keys

Returns

TRUE if dataset has been already pulled, else FALSE

Method get_colnames()

Derive the column names

Usage

MAETealDataset$get_colnames()

Returns

character vector.

Method get_column_labels()

Derive the column labels

Usage

MAETealDataset$get_column_labels()

Returns

character vector.

Method get_ncol()

Get the number of columns of the data

Usage

MAETealDataset$get_ncol()

Returns

numeric vector

Method get_nrow()

Get the number of rows of the data

Usage

MAETealDataset$get_nrow()

Returns

numeric vector

Method get_rownames()

Derive the row names

Usage

MAETealDataset$get_rownames()

Returns

character vector.

Method print()

Prints this MAETealDataset.

Usage

MAETealDataset$print(...)

Arguments

...

additional arguments to the printing method

Returns

invisibly self

Method clone()

The objects of this class are cloneable with this method.

Usage

MAETealDataset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.