A CallablePythonCode class of objects

Super class

teal.data::CodeClass -> PythonCodeClass

Methods

Public methods

PythonCodeClass$eval()

PythonCodeClass$clone()

Inherited methods

teal.data::CodeClass$append()

teal.data::CodeClass$get_code()

teal.data::CodeClass$initialize()

teal.data::CodeClass$set_code()

Method eval()

Evaluates internal code within environment

Usage

PythonCodeClass$eval(
  vars = list(),
  dataname = NULL,
  envir = new.env(parent = parent.env(.GlobalEnv))
)

Arguments

vars

(named list) additional pre-requisite vars to execute code

dataname

(character) name of the data frame object to be returned

envir

(environment) environment in which code will be evaluated

Returns

data.frame containing the mutated dataset

Method clone()

The objects of this class are cloneable with this method.

Usage

PythonCodeClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.