Test if two objects are (nearly) equal

all.equal(target, current) is a utility to compare join_keys objects target and current testing near equality.

Usage

# S3 method for class 'equal.join_keys'
all(target, current, ...)

Arguments

target

R object.

current

other R object, to be compared with target.

...

further arguments for different methods. Not used with join_keys.

Details

If they are different, comparison is still made to some extent, and a report of the differences is returned. Do not use all.equal directly in if expressions—either use isTRUE(all.equal(....)) or identical if appropriate.

The parents attribute comparison tolerates NULL and empty lists and will find no difference.

The list containing all the relationships is treated like a map and ignores entries with NULL if they exist.

See also