Load CDISCTealData object from a file

Usage

cdisc_data_file(path, code = get_code(path))

Arguments

path

A (connection) or a (character)
string giving the pathname of the file or URL to read from. "" indicates the connection stdin.

code

(character)
reproducible code to re-create object

Value

CDISCTealData object

Examples

file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(scda)

     # code>
     ADSL <- synthetic_cdisc_data('latest')$adsl
     ADTTE <- synthetic_cdisc_data('latest')$adtte

     cdisc_data(
          cdisc_dataset(\"ADSL\", ADSL), cdisc_dataset(\"ADTTE\", ADTTE),
          code = \"ADSL <- synthetic_cdisc_data('latest')$adsl
                  ADTTE <- synthetic_cdisc_data('latest')$adtte\",
          check = FALSE
     )
     # <code"
  ),
  con = file_example
)

cdisc_data_file(file_example)
#> A CDISCTealData object containing 2 TealDataset/TealDatasetConnector object(s) as element(s):
#> --> Element 1:
#> A CDISCTealDataset object containing the following data.frame (400 rows and 44 columns):
#>   STUDYID               USUBJID SUBJID SITEID AGE  AGEU SEX
#> 1 AB12345  AB12345-CHN-3-id-128 id-128  CHN-3  32 YEARS   M
#> 2 AB12345 AB12345-CHN-15-id-262 id-262 CHN-15  35 YEARS   M
#> 3 AB12345  AB12345-RUS-3-id-378 id-378  RUS-3  30 YEARS   F
#> 4 AB12345 AB12345-CHN-11-id-220 id-220 CHN-11  26 YEARS   F
#> 5 AB12345  AB12345-CHN-7-id-267 id-267  CHN-7  40 YEARS   M
#> 6 AB12345 AB12345-CHN-15-id-201 id-201 CHN-15  49 YEARS   M
#>                        RACE                 ETHNIC COUNTRY DTHFL         INVID
#> 1                     ASIAN NOT HISPANIC OR LATINO     CHN     N  INV ID CHN-3
#> 2 BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     CHN     N INV ID CHN-15
#> 3                     ASIAN NOT HISPANIC OR LATINO     RUS     N  INV ID RUS-3
#> 4                     ASIAN NOT HISPANIC OR LATINO     CHN     N INV ID CHN-11
#> 5                     ASIAN                UNKNOWN     CHN     N  INV ID CHN-7
#> 6                     ASIAN NOT HISPANIC OR LATINO     CHN     N INV ID CHN-15
#>           INVNAM            ARM ARMCD         ACTARM ACTARMCD         TRT01P
#> 1  Dr. CHN-3 Doe      A: Drug X ARM A      A: Drug X    ARM A      A: Drug X
#> 2 Dr. CHN-15 Doe C: Combination ARM C C: Combination    ARM C C: Combination
#> 3  Dr. RUS-3 Doe C: Combination ARM C C: Combination    ARM C C: Combination
#> 4 Dr. CHN-11 Doe     B: Placebo ARM B     B: Placebo    ARM B     B: Placebo
#> 5  Dr. CHN-7 Doe     B: Placebo ARM B     B: Placebo    ARM B     B: Placebo
#> 6 Dr. CHN-15 Doe C: Combination ARM C C: Combination    ARM C C: Combination
#>           TRT01A REGION1 STRATA1 STRATA2    BMRKR1 BMRKR2 ITTFL SAFFL BMEASIFL
#> 1      A: Drug X    Asia       C      S2 14.424934 MEDIUM     Y     Y        Y
#> 2 C: Combination    Asia       C      S1  4.055463    LOW     Y     Y        N
#> 3 C: Combination Eurasia       A      S1  2.803240   HIGH     Y     Y        Y
#> 4     B: Placebo    Asia       B      S2 10.262734 MEDIUM     Y     Y        Y
#> 5     B: Placebo    Asia       C      S1  6.206763    LOW     Y     Y        N
#> 6 C: Combination    Asia       C      S2  6.906799 MEDIUM     Y     Y        Y
#>   BEP01FL     RANDDT             TRTSDTM             TRTEDTM       EOSSTT
#> 1       Y 2019-02-22 2019-02-24 11:09:18 2021-02-23 22:47:42    COMPLETED
#> 2       N 2019-02-26 2019-02-26 09:05:00 2021-02-25 20:43:24    COMPLETED
#> 3       N 2019-02-24 2019-02-28 03:19:08 2021-02-27 14:57:32    COMPLETED
#> 4       Y 2019-02-27 2019-03-01 13:33:03 2021-03-01 01:11:27    COMPLETED
#> 5       N 2019-03-01 2019-03-02 00:09:16 2021-03-01 11:47:40    COMPLETED
#> 6       N 2019-03-05 2019-03-05 15:23:44 2021-02-17 20:24:57 DISCONTINUED
#>         EOTSTT      EOSDT EOSDY          DCSREAS DTHDT DTHCAUS DTHCAT LDDTHELD
#> 1    COMPLETED 2021-02-23   731             <NA>  <NA>    <NA>   <NA>       NA
#> 2    COMPLETED 2021-02-25   731             <NA>  <NA>    <NA>   <NA>       NA
#> 3    COMPLETED 2021-02-27   731             <NA>  <NA>    <NA>   <NA>       NA
#> 4    COMPLETED 2021-03-01   731             <NA>  <NA>    <NA>   <NA>       NA
#> 5    COMPLETED 2021-03-01   731             <NA>  <NA>    <NA>   <NA>       NA
#> 6 DISCONTINUED 2021-02-17   716 LACK OF EFFICACY  <NA>    <NA>   <NA>       NA
#>   LDDTHGR1   LSTALVDT DTHADY study_duration_secs
#> 1     <NA> 2021-03-05     NA            63113904
#> 2     <NA> 2021-03-15     NA            63113904
#> 3     <NA> 2021-03-15     NA            63113904
#> 4     <NA> 2021-03-17     NA            63113904
#> 5     <NA> 2021-03-25     NA            63113904
#> 6     <NA> 2021-03-01     NA            63113904
#> 
#> ...
#> # A tibble: 6 × 44
#>   STUDYID USUBJID     SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>   <chr>   <chr>       <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#> 1 AB12345 AB12345-CH… id-11  CHN-9     28 YEARS F     NATI… HISPA… CHN     N    
#> 2 AB12345 AB12345-CH… id-352 CHN-16    28 YEARS M     ASIAN HISPA… CHN     N    
#> 3 AB12345 AB12345-CH… id-186 CHN-1     27 YEARS M     ASIAN NOT H… CHN     N    
#> 4 AB12345 AB12345-CH… id-371 CHN-1     28 YEARS F     ASIAN NOT H… CHN     Y    
#> 5 AB12345 AB12345-CH… id-233 CHN-1     36 YEARS F     BLAC… NOT H… CHN     N    
#> 6 AB12345 AB12345-US… id-131 USA-12    44 YEARS F     AMER… NOT H… USA     N    
#> # … with 33 more variables: INVID <chr>, INVNAM <chr>, ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>, TRT01A <fct>, REGION1 <fct>,
#> #   STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <fct>,
#> #   SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>, RANDDT <date>, TRTSDTM <dttm>,
#> #   TRTEDTM <dttm>, EOSSTT <fct>, EOTSTT <fct>, EOSDT <date>, EOSDY <int>,
#> #   DCSREAS <fct>, DTHDT <date>, DTHCAUS <fct>, DTHCAT <fct>, LDDTHELD <int>,
#> #   LDDTHGR1 <fct>, LSTALVDT <date>, DTHADY <int>, study_duration_secs <dbl>
#> --> Element 2:
#> A CDISCTealDataset object containing the following data.frame (1600 rows and 55 columns):
#>   STUDYID              USUBJID SUBJID SITEID AGE  AGEU SEX
#> 1 AB12345 AB12345-BRA-1-id-105 id-105  BRA-1  38 YEARS   M
#> 2 AB12345 AB12345-BRA-1-id-105 id-105  BRA-1  38 YEARS   M
#> 3 AB12345 AB12345-BRA-1-id-105 id-105  BRA-1  38 YEARS   M
#> 4 AB12345 AB12345-BRA-1-id-105 id-105  BRA-1  38 YEARS   M
#> 5 AB12345 AB12345-BRA-1-id-134 id-134  BRA-1  47 YEARS   M
#> 6 AB12345 AB12345-BRA-1-id-134 id-134  BRA-1  47 YEARS   M
#>                        RACE                 ETHNIC COUNTRY DTHFL        INVID
#> 1 BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     BRA     N INV ID BRA-1
#> 2 BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     BRA     N INV ID BRA-1
#> 3 BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     BRA     N INV ID BRA-1
#> 4 BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     BRA     N INV ID BRA-1
#> 5                     WHITE NOT HISPANIC OR LATINO     BRA     N INV ID BRA-1
#> 6                     WHITE NOT HISPANIC OR LATINO     BRA     N INV ID BRA-1
#>          INVNAM       ARM ARMCD    ACTARM ACTARMCD    TRT01P    TRT01A
#> 1 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A A: Drug X A: Drug X
#> 2 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A A: Drug X A: Drug X
#> 3 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A A: Drug X A: Drug X
#> 4 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A A: Drug X A: Drug X
#> 5 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A A: Drug X A: Drug X
#> 6 Dr. BRA-1 Doe A: Drug X ARM A A: Drug X    ARM A A: Drug X A: Drug X
#>         REGION1 STRATA1 STRATA2   BMRKR1 BMRKR2 ITTFL SAFFL BMEASIFL BEP01FL
#> 1 South America       B      S1 4.156914 MEDIUM     Y     Y        Y       Y
#> 2 South America       B      S1 4.156914 MEDIUM     Y     Y        Y       Y
#> 3 South America       B      S1 4.156914 MEDIUM     Y     Y        Y       Y
#> 4 South America       B      S1 4.156914 MEDIUM     Y     Y        Y       Y
#> 5 South America       B      S2 6.462991    LOW     Y     Y        Y       N
#> 6 South America       B      S2 6.462991    LOW     Y     Y        Y       N
#>       RANDDT             TRTSDTM             TRTEDTM       EOSSTT       EOTSTT
#> 1 2020-03-08 2020-03-08 05:28:11 2021-02-17 18:49:59 DISCONTINUED DISCONTINUED
#> 2 2020-03-08 2020-03-08 05:28:11 2021-02-17 18:49:59 DISCONTINUED DISCONTINUED
#> 3 2020-03-08 2020-03-08 05:28:11 2021-02-17 18:49:59 DISCONTINUED DISCONTINUED
#> 4 2020-03-08 2020-03-08 05:28:11 2021-02-17 18:49:59 DISCONTINUED DISCONTINUED
#> 5 2020-11-03 2020-11-04 03:50:33 2022-11-04 15:28:57    COMPLETED    COMPLETED
#> 6 2020-11-03 2020-11-04 03:50:33 2022-11-04 15:28:57    COMPLETED    COMPLETED
#>        EOSDT EOSDY       DCSREAS DTHDT DTHCAUS DTHCAT LDDTHELD LDDTHGR1
#> 1 2021-02-17   347 ADVERSE EVENT  <NA>    <NA>   <NA>       NA     <NA>
#> 2 2021-02-17   347 ADVERSE EVENT  <NA>    <NA>   <NA>       NA     <NA>
#> 3 2021-02-17   347 ADVERSE EVENT  <NA>    <NA>   <NA>       NA     <NA>
#> 4 2021-02-17   347 ADVERSE EVENT  <NA>    <NA>   <NA>       NA     <NA>
#> 5 2022-11-04   731          <NA>  <NA>    <NA>   <NA>       NA     <NA>
#> 6 2022-11-04   731          <NA>  <NA>    <NA>   <NA>       NA     <NA>
#>     LSTALVDT DTHADY study_duration_secs ASEQ TTESEQ
#> 1 2021-03-07     NA            63113904    2      2
#> 2 2021-03-07     NA            63113904    4      4
#> 3 2021-03-07     NA            63113904    3      3
#> 4 2021-03-07     NA            63113904    1      1
#> 5 2022-11-15     NA            63113904    4      4
#> 6 2022-11-15     NA            63113904    2      2
#>                            PARAM PARAMCD      AVAL AVALU       ADTM ADY CNSR
#> 1 Duration of Confirmed Response    CRSD 1273.8498  DAYS 2020-08-13 158    1
#> 2            Event Free Survival     EFS  135.9787  DAYS 2021-02-16 345    0
#> 3               Overall Survival      OS 1759.8678  DAYS 2020-11-05 242    0
#> 4      Progression Free Survival     PFS  241.1912  DAYS 2020-07-23 137    0
#> 5 Duration of Confirmed Response    CRSD  544.5728  DAYS 2021-10-25 355    0
#> 6            Event Free Survival     EFS  401.2725  DAYS 2021-05-05 182    1
#>                      EVNTDESC                      CNSDTDSC
#> 1 Last Date Known To Be Alive    End of AE Reporting Period
#> 2                       Death                              
#> 3                       Death                              
#> 4         Disease Progression                              
#> 5                       Death                              
#> 6 Last Date Known To Be Alive Completion or Discontinuation
#> 
#> ...
#> # A tibble: 6 × 55
#>   STUDYID USUBJID     SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>   <chr>   <chr>       <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#> 1 AB12345 AB12345-US… id-206 USA-8     34 YEARS F     ASIAN NOT H… USA     N    
#> 2 AB12345 AB12345-US… id-206 USA-8     34 YEARS F     ASIAN NOT H… USA     N    
#> 3 AB12345 AB12345-US… id-130 USA-9     40 YEARS M     ASIAN NOT H… USA     N    
#> 4 AB12345 AB12345-US… id-130 USA-9     40 YEARS M     ASIAN NOT H… USA     N    
#> 5 AB12345 AB12345-US… id-130 USA-9     40 YEARS M     ASIAN NOT H… USA     N    
#> 6 AB12345 AB12345-US… id-130 USA-9     40 YEARS M     ASIAN NOT H… USA     N    
#> # … with 44 more variables: INVID <chr>, INVNAM <chr>, ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>, TRT01A <fct>, REGION1 <fct>,
#> #   STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <fct>,
#> #   SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>, RANDDT <date>, TRTSDTM <dttm>,
#> #   TRTEDTM <dttm>, EOSSTT <fct>, EOTSTT <fct>, EOSDT <date>, EOSDY <int>,
#> #   DCSREAS <fct>, DTHDT <date>, DTHCAUS <fct>, DTHCAT <fct>, LDDTHELD <int>,
#> #   LDDTHGR1 <fct>, LSTALVDT <date>, DTHADY <int>, study_duration_secs <dbl>, …