scda TealDatasetConnector

scda_dataset_connector(
  dataname,
  scda_dataname = tolower(dataname),
  scda_name = "latest",
  keys = character(0),
  label = character(0),
  code = character(0),
  script = character(0),
  metadata = list(type = "scda", version = scda_name)
)

scda_cdisc_dataset_connector(
  dataname,
  scda_dataname = tolower(dataname),
  scda_name = "latest",
  keys = get_cdisc_keys(dataname),
  parent = if (identical(dataname, "ADSL")) character(0L) else "ADSL",
  label = character(0),
  code = character(0),
  script = character(0),
  metadata = list(type = "scda", version = scda_name)
)

Arguments

dataname

(character)
A given name for the dataset it may not contain spaces

scda_dataname

(character) which scda dataset to use (e.g. adsl).

scda_name

(character) which version of scda data to take, default "latest".

keys

optional, (character)
vector of dataset primary keys column names

label

(character)
Label to describe the dataset.

code

(character)
A character string defining code to modify raw_data from this dataset. To modify current dataset code should contain at least one assignment to object defined in dataname argument. For example if dataname = ADSL example code should contain ADSL <- <some R code>. Can't be used simultaneously with script

script

(character)
Alternatively to code - location of the file containing modification code. Can't be used simultaneously with script.

metadata

(named list, NULL or CallableFunction)
Field containing either the metadata about the dataset (each element of the list should be atomic and length one) or a CallableFuntion to pull the metadata from a connection. This should return a list or an object which can be converted to a list with as.list.

parent

optional, (character)
parent dataset name

Details

Create a TealDatasetConnector for dataset in scda

Create a CDISCTealDatasetConnector from scda data

Examples

library(scda)
x <- scda_dataset_connector(
  dataname = "ADSL", scda_dataname = "adsl",
)
x$get_code()
#> [1] "ADSL <- scda::synthetic_cdisc_dataset(dataset_name = \"adsl\", name = \"latest\")"
load_dataset(x)
get_dataset(x)
#> A TealDataset object containing the following data.frame (400 rows and 39 columns):
#>   STUDYID               USUBJID SUBJID SITEID AGE SEX                      RACE
#> 1 AB12345  AB12345-CHN-3-id-128 id-128  CHN-3  32   M                     ASIAN
#> 2 AB12345 AB12345-CHN-15-id-262 id-262 CHN-15  35   M BLACK OR AFRICAN AMERICAN
#> 3 AB12345  AB12345-RUS-3-id-378 id-378  RUS-3  30   F                     ASIAN
#> 4 AB12345 AB12345-CHN-11-id-220 id-220 CHN-11  26   F                     ASIAN
#> 5 AB12345  AB12345-CHN-7-id-267 id-267  CHN-7  40   M                     ASIAN
#> 6 AB12345 AB12345-CHN-15-id-201 id-201 CHN-15  49   M                     ASIAN
#>   COUNTRY DTHFL         INVID         INVNAM            ARM ARMCD
#> 1     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe      A: Drug X ARM A
#> 2     CHN     N INV ID CHN-15 Dr. CHN-15 Doe C: Combination ARM C
#> 3     RUS     N  INV ID RUS-3  Dr. RUS-3 Doe C: Combination ARM C
#> 4     CHN     N INV ID CHN-11 Dr. CHN-11 Doe     B: Placebo ARM B
#> 5     CHN     N  INV ID CHN-7  Dr. CHN-7 Doe     B: Placebo ARM B
#> 6     CHN     N INV ID CHN-15 Dr. CHN-15 Doe C: Combination ARM C
#>           ACTARM ACTARMCD REGION1 STRATA1 STRATA2    BMRKR1 BMRKR2 ITTFL SAFFL
#> 1      A: Drug X    ARM A    Asia       C      S2 14.424934 MEDIUM     Y     Y
#> 2 C: Combination    ARM C    Asia       C      S1  4.055463    LOW     Y     Y
#> 3 C: Combination    ARM C Eurasia       A      S1  2.803240   HIGH     Y     Y
#> 4     B: Placebo    ARM B    Asia       B      S2 10.262734 MEDIUM     Y     Y
#> 5     B: Placebo    ARM B    Asia       C      S1  6.206763    LOW     Y     Y
#> 6 C: Combination    ARM C    Asia       C      S2  6.906799 MEDIUM     Y     Y
#>   BMEASIFL BEP01FL     RANDDT             TRTSDTM             TRTEDTM
#> 1        Y       Y 2019-02-22 2019-02-24 11:09:18 2021-02-23 22:47:42
#> 2        N       N 2019-02-26 2019-02-26 09:05:00 2021-02-25 20:43:24
#> 3        Y       N 2019-02-24 2019-02-28 03:19:08 2021-02-27 14:57:32
#> 4        Y       Y 2019-02-27 2019-03-01 13:33:03 2021-03-01 01:11:27
#> 5        N       N 2019-03-01 2019-03-02 00:09:16 2021-03-01 11:47:40
#> 6        Y       N 2019-03-05 2019-03-05 15:23:44 2021-02-17 20:23:53
#>         EOSSTT       EOTSTT      EOSDT EOSDY          DCSREAS DTHDT DTHCAUS
#> 1    COMPLETED    COMPLETED 2021-02-23   731             <NA>  <NA>    <NA>
#> 2    COMPLETED    COMPLETED 2021-02-25   731             <NA>  <NA>    <NA>
#> 3    COMPLETED    COMPLETED 2021-02-27   731             <NA>  <NA>    <NA>
#> 4    COMPLETED    COMPLETED 2021-03-01   731             <NA>  <NA>    <NA>
#> 5    COMPLETED    COMPLETED 2021-03-01   731             <NA>  <NA>    <NA>
#> 6 DISCONTINUED DISCONTINUED 2021-02-17   716 LACK OF EFFICACY  <NA>    <NA>
#>   DTHCAT LDDTHELD LDDTHGR1   LSTALVDT study_duration_secs
#> 1   <NA>       NA     <NA> 2021-03-05            63113904
#> 2   <NA>       NA     <NA> 2021-03-15            63113904
#> 3   <NA>       NA     <NA> 2021-03-15            63113904
#> 4   <NA>       NA     <NA> 2021-03-17            63113904
#> 5   <NA>       NA     <NA> 2021-03-25            63113904
#> 6   <NA>       NA     <NA> 2021-03-01            63113904
#> 
#> ...
#> # A tibble: 6 × 39
#>   STUDYID USUBJID     SUBJID SITEID   AGE SEX   RACE  COUNTRY DTHFL INVID INVNAM
#>   <chr>   <chr>       <chr>  <chr>  <int> <fct> <fct> <fct>   <fct> <chr> <chr> 
#> 1 AB12345 AB12345-CH… id-11  CHN-9     28 F     NATI… CHN     N     INV … Dr. C…
#> 2 AB12345 AB12345-CH… id-352 CHN-16    28 M     ASIAN CHN     N     INV … Dr. C…
#> 3 AB12345 AB12345-CH… id-186 CHN-1     27 M     ASIAN CHN     N     INV … Dr. C…
#> 4 AB12345 AB12345-CH… id-371 CHN-1     28 F     ASIAN CHN     Y     INV … Dr. C…
#> 5 AB12345 AB12345-CH… id-233 CHN-1     36 F     BLAC… CHN     N     INV … Dr. C…
#> 6 AB12345 AB12345-US… id-131 USA-12    44 F     AMER… USA     N     INV … Dr. U…
#> # … with 28 more variables: ARM <fct>, ARMCD <fct>, ACTARM <fct>,
#> #   ACTARMCD <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>, EOSSTT <fct>, EOTSTT <fct>,
#> #   EOSDT <date>, EOSDY <int>, DCSREAS <fct>, DTHDT <date>, DTHCAUS <fct>,
#> #   DTHCAT <fct>, LDDTHELD <int>, LDDTHGR1 <fct>, LSTALVDT <date>,
#> #   study_duration_secs <dbl>
get_dataset(x)$get_metadata()
#> $type
#> [1] "scda"
#> 
#> $version
#> [1] "latest"
#> 
x$get_raw_data()
#> # A tibble: 400 × 39
#>    STUDYID USUBJID    SUBJID SITEID   AGE SEX   RACE  COUNTRY DTHFL INVID INVNAM
#>    <chr>   <chr>      <chr>  <chr>  <int> <fct> <fct> <fct>   <fct> <chr> <chr> 
#>  1 AB12345 AB12345-C… id-128 CHN-3     32 M     ASIAN CHN     N     INV … Dr. C…
#>  2 AB12345 AB12345-C… id-262 CHN-15    35 M     BLAC… CHN     N     INV … Dr. C…
#>  3 AB12345 AB12345-R… id-378 RUS-3     30 F     ASIAN RUS     N     INV … Dr. R…
#>  4 AB12345 AB12345-C… id-220 CHN-11    26 F     ASIAN CHN     N     INV … Dr. C…
#>  5 AB12345 AB12345-C… id-267 CHN-7     40 M     ASIAN CHN     N     INV … Dr. C…
#>  6 AB12345 AB12345-C… id-201 CHN-15    49 M     ASIAN CHN     N     INV … Dr. C…
#>  7 AB12345 AB12345-U… id-45  USA-1     34 F     ASIAN USA     N     INV … Dr. U…
#>  8 AB12345 AB12345-U… id-261 USA-1     32 F     ASIAN USA     N     INV … Dr. U…
#>  9 AB12345 AB12345-N… id-173 NGA-11    24 F     BLAC… NGA     N     INV … Dr. N…
#> 10 AB12345 AB12345-C… id-307 CHN-1     24 M     ASIAN CHN     N     INV … Dr. C…
#> # … with 390 more rows, and 28 more variables: ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>,
#> #   BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>,
#> #   BEP01FL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>, EOSSTT <fct>,
#> #   EOTSTT <fct>, EOSDT <date>, EOSDY <int>, DCSREAS <fct>, DTHDT <date>,
#> #   DTHCAUS <fct>, DTHCAT <fct>, LDDTHELD <int>, LDDTHGR1 <fct>,
#> #   LSTALVDT <date>, study_duration_secs <dbl>

metadata_fun <- callable_function(function(a) list(type = a))
metadata_fun$set_args(args = list(a = "scda"))
y <- scda_dataset_connector(
  dataname = "ADSL", scda_dataname = "adsl",
  metadata = metadata_fun
)
load_dataset(y)
get_dataset(y)$get_metadata()
#> $type
#> [1] "scda"
#>