teal.data: Reproducible data model for teal applications

This package extends teal applications by module which stores the data with they relationships (keys) and reproducible code. Package offers also to load data from files, databases and it's easily extendable by another data sources.

See also

Author

Maintainer: Dawid Kaledkowski dawid.kaledkowski@roche.com (ORCID)

Authors:

Aleksander Chlebowski aleksander.chlebowski@contractors.roche.com (ORCID)

Marcin Kosinski marcin.kosinski.mk1@roche.com

Andre Verissimo andre.verissimo@roche.com (ORCID)

Pawel Rucki pawel.rucki@roche.com

Mahmoud Hallal mahmoud.hallal@roche.com

Nikolas Burkoff

Maciej Nasinski

Konrad Pagacz

Junlue Zhao

Other contributors:

Chendi Liao chendi.liao@roche.com [reviewer]

Dony Unardi unardid@gene.com [reviewer]

F. Hoffmann-La Roche AG [copyright holder, funder]