S3 generic for to_relational_data function.

This function takes an object and converts into a TealData object, the primary data object for use in teal applications.

Usage

to_relational_data(data)

Arguments

data

TealDataset, TealDatasetConnector, data.frame, MultiAssayExperiment, list or function returning a named list.

Value

TealData object

Details

Passing a TealData into this function leaves the object unchanged.

Examples

to_relational_data(head(iris))
#> A TealData object containing 1 TealDataset/TealDatasetConnector object(s) as element(s):
#> --> Element 1:
#> A TealDataset object containing the following data.frame (6 rows and 5 columns):
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa
to_relational_data(dataset("IRIS", head(iris)))
#> A TealData object containing 1 TealDataset/TealDatasetConnector object(s) as element(s):
#> --> Element 1:
#> A TealDataset object containing the following data.frame (6 rows and 5 columns):
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa
to_relational_data(list(iris = head(iris), mtcars = head(mtcars)))
#> A TealData object containing 2 TealDataset/TealDatasetConnector object(s) as element(s):
#> --> Element 1:
#> A TealDataset object containing the following data.frame (6 rows and 5 columns):
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa
#> --> Element 2:
#> A TealDataset object containing the following data.frame (6 rows and 11 columns):
#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> Valiant           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1

d_connector <- dataset_connector("iris", callable_function(function() head(iris)))
d_connector$pull()
to_relational_data(d_connector)
#> A TealData object containing 1 TealDataset/TealDatasetConnector object(s) as element(s):
#> --> Element 1:
#> A TealDatasetConnector object, named iris, containing a TealDataset object that has been loaded/pulled:
#> A TealDataset object containing the following data.frame (6 rows and 5 columns):
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa