Comprehensive data integration function for teal applications

Universal function to pass data to teal application.

Usage

teal_data(..., join_keys = teal.data::join_keys(), code = character(0), check)

Arguments

...

any number of objects (presumably data objects) provided as name = value pairs.

join_keys

(join_keys or single join_key_set) optional object with datasets column names used for joining. If empty then no joins between pairs of objects.

code

(character, language) optional code to reproduce the datasets provided in .... Note this code is not executed and the teal_data may not be reproducible

check

(logical) Use verify() to verify code reproducibility .

Value

A teal_data object.

Examples

teal_data(x1 = iris, x2 = mtcars)
#> ✖ unverified teal_data object
#> <environment: 0x562771d22320> [L]
#> Parent: <environment: devtools_shims>
#> Bindings:
#> • x1: <df[,5]> [L]
#> • x2: <df[,11]> [L]